Get Management Commands Working
This will probably break tests until the api client is updated to merge in the change from dashes to underscores in the package name.

Modified management command and associated methods. Appears to work with a single arbitrarily updated organization (there were no others recently modified).

Discovered that the API calls break if there are more than 500 records, so awaiting modification of the client library to fix that. However, everything on this end seems sound.