Country Code Restoration
Used pycountry package to write a one-off migration script to restore ISO country codes that
used to be pulled from salesforce.

The problem is the MemberSuite Organizations have wildly inconsistent countries (some use the
proper name, some use the abbreviation, etc), so this doesn't fully work. We need to make the
data on MemberSuite more consistent rather than try to account for every possible corner case
in the code here.